refactor: normalize ChatContainer component
