--- web/components/chat/ChatContainer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ChatContainer';
