+++ web/components/chat/index.ts
@@ -1,3 +1,2 @@
-export { default as ChatContainer } from './ChatContainer';
 export { default as ChatUserMessage } from './ChatUserMessage';
 export { default as ChatTextField } from './ChatTextField/ChatTextField';
